pyproject.toml
[tool.poetry]
name = "dbt-remote-state"
version = "0.1.0"
description = "A dbt-core plugin for working with remote state."
authors = ["Nicholas Yager <yager@nicholasyager.com>", "Morgan Kerle <mdgk95@proton.me>"]
readme = "README.md"
packages = [{ include = "dbt_remote_state" }]

[tool.commitizen]
version = "0.1.0"
version_files = ["pyproject.toml:^version"]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
dbt-core = ">=1.6.0,<1.11.0"
requests = "^2.31.0"
google-cloud-storage = "^2.13.0"
boto3 = "^1.28.84"
azure-storage-blob = "^12.19.0"
azure-identity = "^1.15.0"
types-pyyaml = "^6.0.12.12"
paradime-io = "^4.11.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.0"
pytest = "^7.4.0"
isort = "^5.12.0"
dbt-duckdb = ">=1.6.0,<1.10.0"
duckdb = ">=0.8.0"
pre-commit = "^3.6.0"
mypy = "^1.8.0"
types-pytz = "^2025.2.0.20250516"

[tool.poetry.extras]
snowflake = ["dbt-snowflake"]

[tool.poetry.group.docs.dependencies]
mkdocs-material = "^9.5.45"
mike = "^2.1.3"

[tool.ruff]
line-length = 88

[tool.isort]
force_grid_wrap = 0 # Resolve conflict with Black
line_length = 88    # Comply with Ruff and Black

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
